requirements-bot.txt
python-telegram-bot==21.6
web3==6.20.1
solana==0.36.9
solders==0.26.0
httpx==0.28.1
requests==2.32.3
urllib3==2.2.3
certifi==2025.8.3
charset-normalizer==3.4.0
idna==3.10
anyio==4.8.0
sniffio==1.3.1
typing-extensions==4.12.2
python-dotenv==1.0.1
